+++ app/src/main/java/com/yourname/expensetracker/domain/model/dashboard/SpendingSummary.kt
@@ -3,8 +3,8 @@ package com.yourname.expensetracker.domain.model.dashboard
 data class SpendingSummary(
     val totalSpent: Double,
     val previousTotalSpent: Double?,
-    val changePercent: Float?,
-    val dailyHistory: List<Float>,
-    val previousDailyHistory: List<Float>,
+    val changePercent: Double?,
+    val dailyHistory: List<Double>,
+    val previousDailyHistory: List<Double>,
     val transactionCount: Int
 )
